+++ net.sourceforge.pmd.eclipse.plugin/src/main/java/net/sourceforge/pmd/eclipse/runtime/properties/impl/ProjectPropertiesImpl.java
@@ -9,6 +9,8 @@ import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.IOException;
 import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
@@ -59,7 +61,7 @@ public class ProjectPropertiesImpl implements IProjectProperties {
     private boolean pmdEnabled;
     private boolean ruleSetStoredInProject;
     private String ruleSetFile;
-    private RuleSets projectRuleSets;
+    private List<RuleSet> projectRuleSets;
     private long projectRuleFileLastModified = 0;
     private IWorkingSet projectWorkingSet;
     private boolean includeDerivedFiles;
@@ -76,9 +78,8 @@ public class ProjectPropertiesImpl implements IProjectProperties {
         super();
         this.project = project;
         this.projectPropertiesManager = projectPropertiesManager;
-        RuleSets ruleSets = new RuleSets();
-        ruleSets.addRuleSet(PMDPlugin.getDefault().getPreferencesManager().getRuleSet());
-        this.projectRuleSets = ruleSets;
+        this.projectRuleSets = new ArrayList<>();
+        this.projectRuleSets.add(PMDPlugin.getDefault().getPreferencesManager().getRuleSet());
         determineBuildPathIncludesExcludes();
     }
 
@@ -144,9 +145,15 @@ public class ProjectPropertiesImpl implements IProjectProperties {
     }
 
     /**
-     * @see net.sourceforge.pmd.eclipse.runtime.properties.IProjectProperties#getProjectRuleSet()
+     * @deprecated Use {@link #getProjectRuleSetList()}
      */
+    @Deprecated
+    @Override
     public RuleSets getProjectRuleSets() throws PropertiesException {
+        return InternalRuleSetUtil.toRuleSets(projectRuleSets);
+    }
+
+    public List<RuleSet> getProjectRuleSetList() throws PropertiesException {
         return projectRuleSets;
     }
 
@@ -154,23 +161,32 @@ public class ProjectPropertiesImpl implements IProjectProperties {
      * @see net.sourceforge.pmd.eclipse.runtime.properties.IProjectProperties#getProjectRuleSet()
      */
     public RuleSet getProjectRuleSet() throws PropertiesException {
-        return projectRuleSets.getAllRuleSets()[0];
+        return projectRuleSets.get(0);
     }
 
     @Override
     public void setProjectRuleSet(RuleSet projectRuleSet) throws PropertiesException {
-        setProjectRuleSets(new RuleSets(projectRuleSet));
+        setProjectRuleSetList(Collections.singletonList(projectRuleSet));
+    }
+
+    /**
+     * @deprecated Use {@link #setProjectRuleSetList(List)}
+     */
+    @Override
+    @Deprecated
+    public void setProjectRuleSets(final RuleSets newProjectRuleSets) throws PropertiesException {
+        setProjectRuleSetList(Arrays.asList(newProjectRuleSets.getAllRuleSets()));
     }
 
     @Override
-    public void setProjectRuleSets(final RuleSets newProjectRuleSets) throws PropertiesException {
+    public void setProjectRuleSetList(List<RuleSet> newProjectRuleSets) throws PropertiesException {
         LOG.debug("Set a rule set for project {}", this.project.getName());
         if (newProjectRuleSets == null) {
             // TODO: NLS
             throw new PropertiesException("Setting a project rule set to null");
         }
 
-        this.setNeedRebuild(!this.projectRuleSets.getAllRules().equals(newProjectRuleSets.getAllRules()));
+        this.setNeedRebuild(!this.projectRuleSets.equals(newProjectRuleSets));
         this.projectRuleSets = newProjectRuleSets;
         if (this.ruleSetStoredInProject) {
             for (File f : getResolvedRuleSetFiles()) {
@@ -377,7 +393,7 @@ public class ProjectPropertiesImpl implements IProjectProperties {
             // the ruleset writer is only capable of writing one ruleset
             RuleSet ruleSet = RuleSetUtil.newEmpty(RuleSetUtil.DEFAULT_RULESET_NAME,
                     RuleSetUtil.DEFAULT_RULESET_DESCRIPTION);
-            for (RuleSet rs : projectRuleSets.getAllRuleSets()) {
+            for (RuleSet rs : projectRuleSets) {
                 ruleSet = RuleSetUtil.addRules(ruleSet, rs.getRules());
                 ruleSet = InternalRuleSetUtil.addFileInclusions(ruleSet, rs.getFileInclusions());
                 ruleSet = InternalRuleSetUtil.addFileExclusions(ruleSet, rs.getFileExclusions());
@@ -470,7 +486,7 @@ public class ProjectPropertiesImpl implements IProjectProperties {
         }
         if (projectRuleSets != null) {
             projectRuleSetName = "";
-            for (RuleSet rs : projectRuleSets.getAllRuleSets()) {
+            for (RuleSet rs : projectRuleSets) {
                 projectRuleSetName += rs.getName() + ",";
             }
         }